Convert file system objects to `long` The lost most significant bit is unlikely to affect anything. It is also a .NET convention to use signed integers for public APIs even though negative values don't necessarily make sense.

+++ TotalImage/dlgImageInfo.cs
@@ -29,7 +29,7 @@ namespace TotalImage
             lstProperties.Columns[1].Width = lstProperties.ClientRectangle.Width - lstProperties.Columns[0].Width;
 
             lstProperties.FindItemWithText("Filename").SubItems[1].Text = mainForm.filename;
-            lstProperties.FindItemWithText("Size").SubItems[1].Text = Settings.CurrentSettings.SizeUnit.FormatSize((ulong)fileInfo.Length, Settings.CurrentSettings.SizeUnit != SizeUnit.Bytes);
+            lstProperties.FindItemWithText("Size").SubItems[1].Text = Settings.CurrentSettings.SizeUnit.FormatSize(fileInfo.Length, Settings.CurrentSettings.SizeUnit != SizeUnit.Bytes);
             lstProperties.FindItemWithText("Created").SubItems[1].Text = fileInfo.CreationTime.ToString();
             lstProperties.FindItemWithText("Modified").SubItems[1].Text = fileInfo.LastWriteTime.ToString();
             lstProperties.FindItemWithText("Accessed").SubItems[1].Text = fileInfo.LastAccessTime.ToString();
@@ -52,10 +52,10 @@ namespace TotalImage
                 string creatorVersion = $"{vhd.Footer.CreatorVersionMajor}.{vhd.Footer.CreatorVersionMinor}";
                 if(creatorVersion != "0.0")
                     lstProperties.FindItemWithText("Creator version").SubItems[1].Text = creatorVersion;
-            }            
-            
+            }
+
             lstProperties.FindItemWithText("File system").SubItems[1].Text = mainForm.image.PartitionTable.Partitions[mainForm.CurrentPartitionIndex].FileSystem.DisplayName;
-            
+
             var volLabel = mainForm.image.PartitionTable.Partitions[mainForm.CurrentPartitionIndex].FileSystem.VolumeLabel;
             if(!string.IsNullOrWhiteSpace(volLabel))
                 lstProperties.FindItemWithText("Volume label").SubItems[1].Text = volLabel;
@@ -78,8 +78,8 @@ namespace TotalImage
 
             lstProperties.FindItemWithText("Files").SubItems[1].Text = mainForm.image.PartitionTable.Partitions[mainForm.CurrentPartitionIndex].FileSystem.RootDirectory.FileCount(true).ToString();
             lstProperties.FindItemWithText("Subdirectories").SubItems[1].Text = mainForm.image.PartitionTable.Partitions[mainForm.CurrentPartitionIndex].FileSystem.RootDirectory.SubdirectoryCount(true).ToString();
-            lstProperties.FindItemWithText("Total storage capacity").SubItems[1].Text = Settings.CurrentSettings.SizeUnit.FormatSize((ulong)mainForm.image.PartitionTable.Partitions[mainForm.CurrentPartitionIndex].Length, Settings.CurrentSettings.SizeUnit != SizeUnit.Bytes);
-            lstProperties.FindItemWithText("Free space").SubItems[1].Text = Settings.CurrentSettings.SizeUnit.FormatSize((ulong)mainForm.image.PartitionTable.Partitions[mainForm.CurrentPartitionIndex].FileSystem.TotalFreeSpace, Settings.CurrentSettings.SizeUnit != SizeUnit.Bytes);
+            lstProperties.FindItemWithText("Total storage capacity").SubItems[1].Text = Settings.CurrentSettings.SizeUnit.FormatSize(mainForm.image.PartitionTable.Partitions[mainForm.CurrentPartitionIndex].Length, Settings.CurrentSettings.SizeUnit != SizeUnit.Bytes);
+            lstProperties.FindItemWithText("Free space").SubItems[1].Text = Settings.CurrentSettings.SizeUnit.FormatSize(mainForm.image.PartitionTable.Partitions[mainForm.CurrentPartitionIndex].FileSystem.TotalFreeSpace, Settings.CurrentSettings.SizeUnit != SizeUnit.Bytes);
 
             if (mainForm.image.PartitionTable.Partitions[mainForm.CurrentPartitionIndex].FileSystem is FileSystems.FAT.FatFileSystem fs)
             {
@@ -111,7 +111,7 @@ namespace TotalImage
                     Buttons =
                     {
                         TaskDialogButton.Cancel,
-                        TaskDialogButton.Continue                
+                        TaskDialogButton.Continue
                     },
                     Icon = TaskDialogIcon.Warning,
                     DefaultButton = TaskDialogButton.Continue,
@@ -152,8 +152,8 @@ namespace TotalImage
                 sw.WriteLine($"-Attributes: {lstProperties.FindItemWithText("Attributes").SubItems[1].Text}");
 
                 //Skip the hashes if they're not done yet
-                if (hashesDone) 
-                { 
+                if (hashesDone)
+                {
                     sw.WriteLine($"-MD5 hash: {lstProperties.FindItemWithText("MD5 hash").SubItems[1].Text}");
                     sw.WriteLine($"-SHA-1 hash: {lstProperties.FindItemWithText("SHA-1 hash").SubItems[1].Text}");
                 }
@@ -179,7 +179,7 @@ namespace TotalImage
                 sw.WriteLine($"-Total storage capacity: {lstProperties.FindItemWithText("Total storage capacity").SubItems[1].Text}");
                 sw.WriteLine($"-Free space: {lstProperties.FindItemWithText("Free space").SubItems[1].Text}");
                 sw.WriteLine($"-Files: {lstProperties.FindItemWithText("Files").SubItems[1].Text}");
-                sw.WriteLine($"-Subdirectories: {lstProperties.FindItemWithText("Subdirectories").SubItems[1].Text}");          
+                sw.WriteLine($"-Subdirectories: {lstProperties.FindItemWithText("Subdirectories").SubItems[1].Text}");
                 sw.WriteLine();
 
                 if (!string.IsNullOrWhiteSpace(txtComment.Text))
